Add extension points (#26)
Previously, `open-browser-git.nvim` would do its best job to find the
commit to permalink and the remote to link to. Over time, I've improved
the behavior for these functions, e.g. #19 uses the most recent _pushed_
commit rather than just `HEAD`. However, this puts me in the classic
awkward position of either having to support every possible behavior
with an option or leave users in the lurch.

Therefore, I've added some extension points: it is now possible to pass
functions that determine which commit and remote to link to.